Moved graph tests to separate file
