+++ tests/test_caching.py
@@ -9,60 +9,6 @@ from alkymi.alkymi import Status
 from alkymi.recipe import Recipe
 
 
-def test_graph() -> None:
-    AlkymiConfig.get().cache = False
-
-    @alk.recipe()
-    def a() -> str:
-        return "a"
-
-    @alk.recipe()
-    def b() -> str:
-        return "b"
-
-    @alk.recipe()
-    def c() -> str:
-        return "c"
-
-    @alk.recipe()
-    def depends_a(a: str) -> List[str]:
-        return [a, a, a, a]
-
-    @alk.foreach(depends_a)
-    def foreach_a(a: str) -> str:
-        return a + "_"
-
-    @alk.recipe()
-    def depends_ab(a: str, b: str) -> str:
-        return a + b
-
-    @alk.recipe()
-    def root(foreach_a: List[str], depends_ab: str, c: str) -> str:
-        return "".join(foreach_a) + depends_ab + c
-
-    graph = alk.alkymi.create_graph(root)
-    # import matplotlib.pyplot as plt
-    # import networkx as nx
-    # plt.subplot(121)
-    # color_state_map = {Status.Ok: 'green', Status.NotEvaluatedYet: 'red', Status.MappedInputsDirty: "yellow"}
-    # nx.draw(graph, with_labels=True, font_weight='bold', node_color=[color_state_map[node[1]['status']]
-    #                                                                  for node in graph.nodes(data=True)])
-    # plt.show()
-    # print(graph)
-    assert len(graph.nodes) > 1
-    assert graph.has_successor(a, depends_a)
-    assert graph.has_successor(a, depends_ab)
-    assert graph.has_successor(b, depends_ab)
-    assert graph.has_successor(depends_a, foreach_a)
-    assert graph.has_successor(foreach_a, root)
-    assert graph.has_successor(depends_ab, root)
-    assert graph.has_successor(c, root)
-
-    result, _ = alk.alkymi.evaluate_recipe(root, graph)
-    assert result == "a_a_a_a_abc"
-
-
-
 def test_caching(caplog, tmpdir):
     """
     Test that a cache is created (in the set location), and that recipe can be restored correctly
